switch up variable types so it runs(temporary)

+++ emulator.h
@@ -8,7 +8,7 @@ typedef struct {
     uint8_t memory[4096];
     uint16_t stack[15];
     uint8_t sp;
-    uint16_t counter;
+    uint8_t counter;
     uint16_t index;
     uint8_t timer;
     uint8_t sound_timer;
